Update modules structure
